style(typings): Update structure
