+++ src/addons/Select/index.d.ts
@@ -1,21 +1,1 @@
-import * as React from 'react';
-import {
-  DropdownDivider,
-  DropdownHeader,
-  DropdownItem,
-  DropdownMenu,
-  DropdownProps
-} from '../../modules/Dropdown';
-
-export interface SelectProps extends DropdownProps {
-  selection: true;
-}
-
-interface SelectComponent extends React.StatelessComponent<SelectProps> {
-  Divider: typeof DropdownDivider;
-  Header: typeof DropdownHeader;
-  Item: typeof DropdownItem;
-  Menu: typeof DropdownMenu;
-}
-
-export const Select: SelectComponent;
+export * from './Select';
