+++ src/collections/Form/FormButton.d.ts
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+import { ButtonProps } from '../../elements/Button';
+import { FormFieldProps } from './FormField';
+
+interface FormButtonProps extends FormFieldProps, ButtonProps {
+}
+
+const FormButton: React.StatelessComponent<FormButtonProps>;
+
+export default FormButton;
+
