+++ src/collections/Form/FormDropdown.d.ts
@@ -1,0 +1,11 @@
+import * as React from 'react';
+
+import { DropdownProps } from '../../modules/Dropdown';
+import { FormFieldProps } from './FormField';
+
+interface FormDropdownProps extends FormFieldProps, DropdownProps {
+}
+
+const FormDropdown: React.StatelessComponent<FormDropdownProps>;
+
+export default FormDropdown;
