+++ src/collections/Form/FormInput.d.ts
@@ -1,0 +1,11 @@
+import * as React from 'react';
+
+import { InputProps } from '../../elements/Input';
+import { FormFieldProps } from './FormField';
+
+interface FormInputProps extends FormFieldProps, InputProps {
+}
+
+const FormInput: React.StatelessComponent<FormInputProps>;
+
+export default FormInput;
