+++ src/collections/Form/FormSelect.d.ts
@@ -1,0 +1,11 @@
+import * as React from 'react';
+
+import { SelectProps } from '../../addons/Select';
+import { FormFieldProps } from './FormField';
+
+interface FormSelectProps extends FormFieldProps, SelectProps {
+}
+
+const FormSelect: React.StatelessComponent<FormSelectProps>;
+
+export default FormSelect;
