+++ src/collections/Form/FormTextArea.d.ts
@@ -1,0 +1,11 @@
+import * as React from 'react';
+
+import { TextAreaProps } from '../../addons/TextArea';
+import { FormFieldProps } from './FormField';
+
+interface FormTextAreaProps extends FormFieldProps, TextAreaProps {
+}
+
+const FormTextArea: React.StatelessComponent<FormTextAreaProps>;
+
+export default FormTextArea;
